getFavRides query

+++ src/main/java/eus/ehu/sharetrip/domain/Ride.java
@@ -137,7 +137,7 @@ public class Ride implements Serializable {
 	}
 
 	
-	public float getNumPlaces() {
+	public int getNumPlaces() {
 		return numPlaces;
 	}
